Feat: lagde et globalt skade system

+++ Assets/Scripts/Player/PlayerStateManager.cs
@@ -31,7 +31,7 @@ public class PlayerStateManager : MonoBehaviour
     public GameObject RightArmCollider;
     public GameObject AirAttackCollider;
 
-    public int CurrentPlayerHealth;
+    public float CurrentPlayerHealth;
 
     [SerializeField] GameObjectScrub PlayerReference;
     private void Start()
